more work on swapchain

+++ core/QueueFamilyFlags.h
@@ -3,7 +3,7 @@
 #include "Bitmask.h"
 
 namespace Vixen {
-    enum class QueueFamilyFlags : int64_t {
+    enum class QueueFamilyFlags : uint32_t {
         Graphics = 1 << 0,
         Transfer = 1 << 1,
         Compute = 1 << 2
