+++ core/image/ImageAspect.h
@@ -3,7 +3,7 @@
 #include "core/Bitmask.h"
 
 namespace Vixen {
-    enum class ImageAspect : int64_t {
+    enum class ImageAspect : uint32_t {
         Color = 1 << 0,
         Depth = 1 << 1,
         Stencil = 1 << 2
